Calculator

+++ src/example_1005/Calculator.java
@@ -22,6 +22,6 @@ public class Calculator {
         return a*b;
     }
     public double divide(){
-        return a/b;
+        return (double)a/(double)b;
     }
 }
